+++ CHANGELOG.md
@@ -2,6 +2,7 @@ medialibrary CHANGELOG
 ======================
 
 ## CURRENT
+ - [medialibrary-19](https://github.com/cjcodeproj/medialibrary/issues/19) Update code for Character Names
  - [medialibrary-16](https://github.com/cjcodeproj/medialibrary/issues/16) FilenameMatches regex pattern is incorrect
  - [medialibrary-15](https://github.com/cjcodeproj/medialibrary/issues/15) The keyword code is kind of sloppy
  - [medialibrary-14](https://github.com/cjcodeproj/medialibrary/issues/14) The proper noun code is kind of sloppy